Use flask instead of node.
translated node server to flask instead.
